+++ CHANGELOG.md
@@ -1,10 +1,9 @@
 ## September 18th 2022
 
 - Added level state initialization and playing.
-
+- Player turn switcher.
 
 TODO: 
-- Player turn switcher.
 - Add Pause button to the level.
 - Create a check mechanics for the word with already written words.
 -- prototype stage